[PR review]: implement PartialEq for Value for readability
